poprawki

+++ src/main/java/com/app/ecommerceapp/model/Product.java
@@ -1,10 +1,12 @@
 package com.app.ecommerceapp.model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private String id;
     private String name;
     private String description;
-    private double price;
-    private int quantity;
+    private BigDecimal price;
+    private Integer quantity;
     private String categoryId;
 }
